requirements.txt
# Core deep learning
tensorflow>=2.13,<2.17

# Computer vision
opencv-python>=4.8

# Image processing
Pillow>=10.0

# Scientific computing
numpy>=1.24
scikit-learn>=1.3

# Visualization & reporting
matplotlib>=3.7
seaborn>=0.13

# Notebook support (optional but recommended)
jupyter>=1.0
ipykernel>=6.0
